Fix int to int64

+++ httpserver/httpserver.go
@@ -42,7 +42,7 @@ type HttpServerImpl struct {
 	ReadHeaderTimeout              time.Duration
 	SenzingEngineConfigurationJson string
 	SenzingModuleName              string
-	SenzingVerboseLogging          int
+	SenzingVerboseLogging          int64
 	ServerAddress                  string
 	ServerOptions                  []senzingrestapi.ServerOption
 	ServerPort                     int
